Calculate (and potentially plot) winrates by character.

Calculate (and potentially plot) winrates by character.

Usage

winrate(
  runhistory,
  plotit = TRUE,
  ci = TRUE,
  samples = 200,
  lower_quant = 0.025,
  upper_quant = 0.975
)

Arguments

runhistory

An STS2RunHistory object, filtered if necessary.

plotit

Whether to plot the data or just return a summary data.frame

ci

Whether to calculate and plot 95% confidence intervals

samples

The number of iterations to use when generating CIs.

lower_quant

The lower quantile of the CIs.

upper_quant

The upper quantile of the CIs.

Value

A data.frame summary of winrates by character.

Examples

if (FALSE) { # interactive()
myruns <- load_sts_history()
winrate(myruns)
}